refactor: simplify CashFlow web hook/page duplication
- Extract currentYearMonth/formatMonthInputValue/parseMonthInputValue
  (byte-for-byte duplicated across 4 hooks: useMonthly, useMensais,
  useControleMae, useInvestmentSnapshots) into utils/formatters.ts.
- Extract the identical isEditing Save/Cancel button pair (duplicated
  across 4 page Row components) into a shared EditRowActions component.
- financialApiClient: collapse requestVoid's duplicated fetch/error-check
  logic into a shared sendRequest, reused by both request<T> and
  requestVoid.
- Remove useMonthly's dead markPaidError state field - never read, and the
  MARK_PAID_ERROR action already correctly writes to saveError instead.
- MonthlyPage: reuse the hook's exported EditField type instead of
  re-declaring the same field-name union inline.

No behavior change - full web test suite still green (490/490), typecheck
and lint clean.
