+++ Financial.Web/src/pages/ControleMaePage.tsx
@@ -1,4 +1,5 @@
 import type { MaeLedgerEntryDto } from '../api/types'
+import EditRowActions from '../components/EditRowActions'
 import ErrorState from '../components/ErrorState'
 import LoadingState from '../components/LoadingState'
 import { useControleMae } from '../hooks/useControleMae'
@@ -51,12 +52,7 @@ function EntryRow({
           />
         </td>
         <td>
-          <button type="button" disabled={isSaving} onClick={onSave}>
-            {isSaving ? 'Saving...' : 'Save'}
-          </button>
-          <button type="button" onClick={onCancel}>
-            Cancel
-          </button>
+          <EditRowActions isSaving={isSaving} onSave={onSave} onCancel={onCancel} />
         </td>
       </tr>
     )
