+++ Financial.Web/src/pages/InvestmentSnapshotsPage.tsx
@@ -1,4 +1,5 @@
 import type { InvestmentSnapshotDto } from '../api/types'
+import EditRowActions from '../components/EditRowActions'
 import ErrorState from '../components/ErrorState'
 import LoadingState from '../components/LoadingState'
 import { useInvestmentSnapshots } from '../hooks/useInvestmentSnapshots'
@@ -36,12 +37,7 @@ function SnapshotRow({
           <input type="number" step="0.01" min="0" value={editValue} onChange={(e) => onValueChange(e.target.value)} />
         </td>
         <td>
-          <button type="button" disabled={isSaving} onClick={onSave}>
-            {isSaving ? 'Saving...' : 'Save'}
-          </button>
-          <button type="button" onClick={onCancel}>
-            Cancel
-          </button>
+          <EditRowActions isSaving={isSaving} onSave={onSave} onCancel={onCancel} />
         </td>
       </tr>
     )
