+++ Financial.Web/src/pages/MonthlyPage.tsx
@@ -1,7 +1,8 @@
 import type { ExpenseDto } from '../api/types'
+import EditRowActions from '../components/EditRowActions'
 import ErrorState from '../components/ErrorState'
 import LoadingState from '../components/LoadingState'
-import { useMonthly } from '../hooks/useMonthly'
+import { useMonthly, type EditField } from '../hooks/useMonthly'
 import { formatN2, formatShortDate } from '../utils/formatters'
 import './MonthlyPage.css'
 
@@ -37,10 +38,7 @@ interface ExpenseRowProps {
   editCardTag: string
   isSaving: boolean
   onEdit: (expense: ExpenseDto) => void
-  onFieldChange: (
-    field: 'editDate' | 'editDescription' | 'editValue' | 'editCategory' | 'editPaymentSource' | 'editCardTag',
-    value: string,
-  ) => void
+  onFieldChange: (field: EditField, value: string) => void
   onSave: () => void
   onCancel: () => void
   onDelete: (id: string) => void
@@ -107,12 +105,7 @@ function ExpenseRow({
           </select>
         </td>
         <td>
-          <button type="button" disabled={isSaving} onClick={onSave}>
-            {isSaving ? 'Saving...' : 'Save'}
-          </button>
-          <button type="button" onClick={onCancel}>
-            Cancel
-          </button>
+          <EditRowActions isSaving={isSaving} onSave={onSave} onCancel={onCancel} />
         </td>
       </tr>
     )
